Add files via upload
Our project 'Fall Detection' represents the novelty with six parameters and a mild alarm,
contributing to fall prevention and improved safety for elderly individuals.To
achieve this firstly gait parameters like human gait and balance: step time,Stride
time,step length,Stride length,step velocity,Step Count on heel and toe is gathered
IMU(Accelerometer,Gyroscope,Magnetometer).The real time gathered data is then
analysed using the pattern recognition algorithm(Random forest). The real time gait
data is checked and based on this the intimation is given the elderly persons. The
alert is planned to be given in the form of physical mode by a small trigger or by a
voice note.Finally all the results of gait parameters and everything is planned to be
showed in real time through app built by React native.